fix(s3-gpio48): 64-bit pin masks throughout

+++ esp32-s3/shims/hardware/gpio.h
@@ -11,7 +11,7 @@
 extern "C" {
 #endif
 
-uint32_t gpio_get_all(void);
+uint64_t gpio_get_all(void);
 
 #ifdef __cplusplus
 }
